Fix/deal with residual rounding issues

+++ src/v4vapp_backend_v2/models/invoice_models.py
@@ -305,7 +305,7 @@ class Invoice(TrackedBaseModel):
 
         self.conv = CryptoConversion(
             conv_from=Currency.MSATS,
-            value=float(amount_msat),
+            value=int(amount_msat),
             quote=quote,
         ).conversion
